HomeScreen now display nearby beacons with the attachments attached to the beacon.
The reporter service has now been updated to hit the new Server URL with the new payload.
Note: IMEI of device is currently hardcoded.

I have not deleted the old code of identifying beacons. However I unimplemented that way of tracking beacons. Instead, the app is using the new Proximity SDK of Estimote.

Server error response are currently not handled. Instead they can be seen through the debugger, but not in the App. Will be working on these later on.