Implementing beacon tracking while app is closed or killed.
Beacon scan will still run if the app is killed.

Change in getting the singleton for beacon implementation.

Need to make recyclerview to update the view for current beacons that can be seen.

Make the notification that keeps the beacon scanning on the background after app is killed show nicer.