Updating notification message.
Sensors such as: IMEI, AndroidID, Mac Address, and networking status are not being disabled or enabled as checkbox is checked or unchecked. This are not sensors, rather than just getters from system services. Their battery consumptions is very low if this are always turned on.

All other sensors should be working as expected.